Crate hoover3_client

The Hoover3 client library - the website frontend components and pages.

Modules

api

This module contains the server wrappers for all the API functions.

app

App component - the whole website as a single component.

components

This module contains the components for the Hoover3 client. Components are reusable UI elements that can be used in multiple pages.

errors

This module contains the error handling for the Hoover3 client. It provides an implementation for the .throw() method for Results and Options,

pages

This module contains the pages for the Hoover3 client. Pages are the main UI elements that are displayed to the user.

routes

This module contains the routes for the Hoover3 client. Routes are the URLs that are used to navigate the client.

time

This module contains time helpers for WASM clients.