hoover3_client
Module api

This module contains the server wrappers for all the API functions.

Modules

__create_datasource__ 🔒

__create_new_collection__ 🔒

__db_explorer_run_query__ 🔒

__drop_collection__ 🔒

__get_all_collections__ 🔒

__get_all_datasources__ 🔒

__get_collection_schema__ 🔒

__get_container_status__ 🔒

__get_datasource__ 🔒

__get_scan_status__ 🔒

__get_single_collection__ 🔒

__get_workflow_status_tree__ 🔒

__list_directory__ 🔒

__scylla_row_count__ 🔒

__start_scan__ 🔒

__update_collection__ 🔒

__wait_for_scan_results__ 🔒

Macros

server_wrapper 🔒

Structs

ServerCallEvent

Struct records previous server calls, their timing and results.

Functions

__create_datasource__

API Client wrapper - server part

__create_new_collection__

API Client wrapper - server part

__db_explorer_run_query__

API Client wrapper - server part

__drop_collection__

API Client wrapper - server part

__get_all_collections__

API Client wrapper - server part

__get_all_datasources__

API Client wrapper - server part

__get_collection_schema__

API Client wrapper - server part

__get_container_status__

API Client wrapper - server part

__get_datasource__

API Client wrapper - server part

__get_scan_status__

API Client wrapper - server part

__get_single_collection__

API Client wrapper - server part

__get_workflow_status_tree__

API Client wrapper - server part

__list_directory__

API Client wrapper - server part

__scylla_row_count__

API Client wrapper - server part

__start_scan__

API Client wrapper - server part

__update_collection__

API Client wrapper - server part

__wait_for_scan_results__

API Client wrapper - server part

_after_call 🔒

Push a server end call event to the server call history. This causes the loading spinner to be hidden if no other calls are running..

_before_call 🔒

Push a server start call event to the server call history. This causes the loading spinner to be shown.

create_datasource

API Client wrapper - client part - for the backend function [hoover3_database::client_query::datasources :: create_datasource] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

create_new_collection

API Client wrapper - client part - for the backend function [hoover3_database::client_query::collections :: create_new_collection] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

db_explorer_run_query

API Client wrapper - client part - for the backend function [hoover3_database::client_query::database_explorer :: db_explorer_run_query] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

drop_collection

API Client wrapper - client part - for the backend function [hoover3_database::client_query::collections :: drop_collection] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

flatten_result

Result::flatten() is unstable/nightly, so we implement it here.

get_all_collections

API Client wrapper - client part - for the backend function [hoover3_database::client_query::collections :: get_all_collections] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

get_all_datasources

API Client wrapper - client part - for the backend function [hoover3_database::client_query::datasources :: get_all_datasources] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

get_collection_schema

API Client wrapper - client part - for the backend function [hoover3_database::migrate :: get_collection_schema] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

get_container_status

API Client wrapper - client part - for the backend function [hoover3_database::client_query::docker_health :: get_container_status] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

get_datasource

API Client wrapper - client part - for the backend function [hoover3_database::client_query::datasources :: get_datasource] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

get_scan_status

API Client wrapper - client part - for the backend function [hoover3_filesystem_scanner :: get_scan_status] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

get_single_collection

API Client wrapper - client part - for the backend function [hoover3_database::client_query::collections :: get_single_collection] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

get_workflow_status_tree

API Client wrapper - client part - for the backend function [hoover3_taskdef::tasks::status_tree :: get_workflow_status_tree] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

list_directory

API Client wrapper - client part - for the backend function [hoover3_database::client_query::list_disk :: list_directory] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

scylla_row_count

API Client wrapper - client part - for the backend function [hoover3_database::client_query::database_explorer :: scylla_row_count] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

start_scan

API Client wrapper - client part - for the backend function [hoover3_filesystem_scanner :: start_scan] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

truncate 🔒

update_collection

API Client wrapper - client part - for the backend function [hoover3_database::client_query::collections :: update_collection] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.

wait_for_scan_results

API Client wrapper - client part - for the backend function [hoover3_filesystem_scanner :: wait_for_scan_results] This is the wrapper function that runs on the client. It retries the server call up to 2 times, and logs the results using _before_call and _after_call.