hoover3_client
Module components

This module contains the components for the Hoover3 client. Components are reusable UI elements that can be used in multiple pages.

Re-exports

pub use DynamicTable_completions::Component::DynamicTable;

pub use DynamicTableInner_completions::Component::DynamicTableInner;

pub use PageTitle_completions::Component::PageTitle;

Modules

table 🔒

titles 🔒

Structs

DynamicTableInnerProps

Properties for the DynamicTableInner component.

DynamicTableProps

Properties for the DynamicTable component.

HtmlTableProps_

Props for the HtmlTable component:

InfoCardProps_

Props for the InfoCard component:

PageTitleProps

Properties for the PageTitle component.

Traits

DataRowDisplay

Trait for displaying a data row in a table. The column names and their types are known at compile time.

Functions

DynamicTable

Table component that displays the result of a dynamic query. The data format is not known at compile time. This specific component also displays metadata like the query execution time.

DynamicTableInner

Table component that displays the result of a dynamic query. The data format is not known at compile time.

HtmlTable

Component that displays a table of data, using the DataRowDisplay trait for rendering. The data is passed in as a vector of rows.

InfoCard

Component that displays a card with a title, fields for a single row, and a form to edit the data if possible.

PageTitle

Component that displays a page title. Level is 0 for no wrapper, 1 for h1, 2 for h2, etc. Subtitle is rendered as a small gray text to the left of the title.

make_page_title

Create a page title element. This is a helper function for use in other components.