hoover3_client::components::table
Trait DataRowDisplay

pub trait DataRowDisplay: Serialize + for<'a> Deserialize<'a> {
    // Provided methods
    fn get_headers() -> Vec<&'static str> { ... }
    fn render_cell(&self, header_name: &str) -> Element { ... }
    fn can_edit(_header_name: &str) -> bool { ... }
    fn get_editable_fields(&self) -> BTreeMap<String, String> { ... }
    fn set_editable_fields(&mut self, _h: BTreeMap<String, String>) { ... }
}

Trait for displaying a data row in a table. The column names and their types are known at compile time.

Provided Methods

fn get_headers() -> Vec<&'static str>

Get the headers of the table. These must be known at compile time.

fn render_cell(&self, header_name: &str) -> Element

Render a cell in the table. Default implementation dumps the row as JSON and reads the column.

fn can_edit(_header_name: &str) -> bool

Check if a column can be edited. These must be known at compile time. Default implementation is false.

fn get_editable_fields(&self) -> BTreeMap<String, String>

Get the values of the editable fields in the row. This is required to populate forms to edit the row.

fn set_editable_fields(&mut self, _h: BTreeMap<String, String>)

Set the editable fields on the row. This function is used to update a row from form values (input and textarea html elements).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.