hoover3_client
Module pages

This module contains the pages for the Hoover3 client. Pages are the main UI elements that are displayed to the user.

Re-exports

pub use HomePage_completions::Component::HomePage;

pub use dx_translate::DioxusTranslatePage;

pub use dx_translate::DioxusTranslatePage;

pub use DashboardIframePage_completions::Component::DashboardIframePage;

pub use DashboardNavbarDropdown_completions::Component::DashboardNavbarDropdown;

pub use DashboardsHomePage_completions::Component::DashboardsHomePage;

pub use docker_health::DockerHealthPage;

pub use docker_health::DockerHealthPage;

pub use server_call_log::ServerCallLogPage;

pub use server_call_log::ServerCallLogPage;

pub use DatabaseExplorerQueryToolPage_completions::Component::DatabaseExplorerQueryToolPage;

pub use DatabaseExplorerPage_completions::Component::DatabaseExplorerPage;

pub use DatabaseExplorerRootPage_completions::Component::DatabaseExplorerRootPage;

pub use ToolsHomePage_completions::Component::ToolsHomePage;

pub use CollectionsAdminListPage_completions::Component::CollectionsAdminListPage;

pub use CollectionAdminDetailsPage_completions::Component::CollectionAdminDetailsPage;

pub use DatasourceAdminDetailsPage_completions::Component::DatasourceAdminDetailsPage;

pub use WorkflowStatusDisplay_completions::Component::WorkflowStatusDisplay;

pub use NewDatasourceFormPage_completions::Component::NewDatasourceFormPage;

pub use AdminHomePage_completions::Component::AdminHomePage;

Modules

admin 🔒

home_page 🔒

tools 🔒

Structs

CollectionAdminDetailsPageProps

Properties for the CollectionAdminDetailsPage component.

DashboardIframePageProps

Properties for the DashboardIframePage component.

DatabaseExplorerPageProps

Properties for the DatabaseExplorerPage component.

DatabaseExplorerQueryToolPageProps

Properties for the DatabaseExplorerQueryToolPage component.

DatasourceAdminDetailsPageProps

Properties for the DatasourceAdminDetailsPage component.

NewDatasourceFormPageProps

Properties for the NewDatasourceFormPage component.

SqlQueryToolState

State for the SQL query tool, the part kept in the URL.

WorkflowStatusDisplayProps

Properties for the WorkflowStatusDisplay component.

Enums

DatabaseExplorerRoute

Router for the database explorer sub-app.

Functions

AdminHomePage

CollectionAdminDetailsPage

Admin Page that displays the details of a collection and a list of data sources.

CollectionsAdminListPage

Admin Page that displays the list of collections.

DashboardIframePage

Page that displays a dashboard iframe.

DashboardNavbarDropdown

Navbar dropdown that displays a list of dashboard links.

DashboardsHomePage

DatabaseExplorerPage

Main Page that displays the database explorer. Because the UrlParam data is not available immediately after the component is mounted, we need to use a signal to delay the rendering of the page until the UrlParam is available.

DatabaseExplorerQueryToolPage

Page that displays the SQL query tool.

DatabaseExplorerRootPage

DatasourceAdminDetailsPage

Admin Page that displays the details of a data source.

DioxusTranslatePage

Page that displays a tool for translating HTML to DX.

DockerHealthPage

Page that displays the health of Docker containers.

HomePage

Home page with main search interface.

NewDatasourceFormPage

Admin Page that displays the form for creating a new data source. It shows a form and a file browser to select the folder for the data source.

ServerCallLogPage

Page that displays the server call log table.

ToolsHomePage

WorkflowStatusDisplay

Component that displays the status of a workflow, including self-refreshing progress bar.