hoover3_client::pages::admin
Module new_datasource_form

Re-exports

pub use NewDatasourceFormPage_completions::Component::NewDatasourceFormPage;

Structs

DatasourcePathPickerProps 🔒

Properties for the DatasourcePathPicker component.

NewDatasourceFormPageProps

Properties for the NewDatasourceFormPage component.

_NewDatasourceFormPageProps 🔒

Properties for the _NewDatasourceFormPage component.

Functions

DatasourcePathPicker 🔒

Props

NewDatasourceFormPage

Admin Page that displays the form for creating a new data source. It shows a form and a file browser to select the folder for the data source.

_NewDatasourceFormPage 🔒

Actual page that displays the form for creating a new data source.

display_file_name 🔒

display_path 🔒

format_time 🔒