hoover3_client::pages::tools
Module database_explorer

Re-exports

pub use DatabaseExplorerPage_completions::Component::DatabaseExplorerPage;

pub use DatabaseExplorerRootPage_completions::Component::DatabaseExplorerRootPage;

pub use sql_query_tool::*;

Modules

sql_query_tool 🔒

Structs

CardGridDisplayProps 🔒

Properties for the CardGridDisplay component.

CollectionStatsInfoCardProps 🔒

Properties for the CollectionStatsInfoCard component.

DatabaseExplorerCollectionPageGraphEdgesProps 🔒

Properties for the DatabaseExplorerCollectionPageGraphEdges component.

DatabaseExplorerCollectionPageGraphNodesProps 🔒

Properties for the DatabaseExplorerCollectionPageGraphNodes component.

DatabaseExplorerCollectionPageProps 🔒

Properties for the DatabaseExplorerCollectionPage component.

DatabaseExplorerCollectionPageSearchIndexProps 🔒

Properties for the DatabaseExplorerCollectionPageSearchIndex component.

DatabaseExplorerCollectionPageSqlTableProps 🔒

Properties for the DatabaseExplorerCollectionPageSqlTable component.

DatabaseExplorerHeaderProps 🔒

Properties for the DatabaseExplorerHeader component.

DatabaseExplorerPageProps

Properties for the DatabaseExplorerPage component.

GraphEdgesCardsProps 🔒

Properties for the GraphEdgesCards component.

GraphNodesCardsProps 🔒

Properties for the GraphNodesCards component.

LinkCardProps 🔒

Properties for the LinkCard component.

SQLRowCountDisplayStringProps 🔒

Properties for the SQLRowCountDisplayString component.

SQLTableCardsProps 🔒

Properties for the SQLTableCards component.

SearchIndexCardsProps 🔒

Properties for the SearchIndexCards component.

_DatabaseExplorerPageProps 🔒

Properties for the _DatabaseExplorerPage component.

Enums

DatabaseExplorerRoute

Router for the database explorer sub-app.

Functions

CardGridDisplay 🔒

Props

CollectionStatsInfoCard 🔒

Props

DatabaseExplorerCollectionPage 🔒

Props

DatabaseExplorerCollectionPageGraphEdges 🔒

Props

DatabaseExplorerCollectionPageGraphNodes 🔒

Props

DatabaseExplorerCollectionPageSearchIndex 🔒

Props

DatabaseExplorerCollectionPageSqlTable 🔒

Props

DatabaseExplorerHeader 🔒

Props

DatabaseExplorerPage

Main Page that displays the database explorer. Because the UrlParam data is not available immediately after the component is mounted, we need to use a signal to delay the rendering of the page until the UrlParam is available.

DatabaseExplorerRootPage

GraphEdgesCards 🔒

Props

GraphNodesCards 🔒

Props

LinkCard 🔒

Props

SQLRowCountDisplayString 🔒

Props

SQLTableCards 🔒

Props

SearchIndexCards 🔒

Props

_DatabaseExplorerPage 🔒

Internal component that switches between the different pages of the database explorer.