hoover3_client
Module routes

This module contains the routes for the Hoover3 client. Routes are the URLs that are used to navigate the client.

Re-exports

pub use NavbarDropdown_completions::Component::NavbarDropdown;

Modules

url_param 🔒

Structs

DisplayErrorProps 🔒

Properties for the DisplayError component.

NavbarDropdownProps

Properties for the NavbarDropdown component.

PageNotFoundProps 🔒

Properties for the PageNotFound component.

UrlParam

Helper for saving structs in URL as base64 strings.

Enums

Route

The enum of all the routes for the Hoover3 client. To nest another route type object inside a page, use UrlParam<T>.

Functions

DisplayError 🔒

Props

Navbar 🔒

NavbarDropdown

Component for a dropdown menu in the navbar.

NavbarLayout 🔒

Component that wraps the main page with the navbar and error handler. Also initializes the component tracking backend server calls.

NavbarTitleCrumbs 🔒

PageNotFound 🔒

Component for displaying a page not found 404 wrong URL error.